# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lonely Bypass API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BypassInput {
  url: string;
  /** @nullable */
  apiKey?: string | null;
}

export interface BypassResult {
  success: boolean;
  destination: string;
  /** @nullable */
  service?: string | null;
  cached?: boolean;
}

export interface BulkBypassInput {
  urls: string[];
  /** @nullable */
  apiKey?: string | null;
}

export interface BulkBypassResult {
  results: BypassResult[];
}

export interface Service {
  id: string;
  name: string;
  domains: string[];
  category: string;
}

export interface Stats {
  totalBypassed: number;
  totalServices: number;
  uptime: string;
}

export interface ApiKeyInput {
  label: string;
}

export interface ApiKey {
  id: string;
  key: string;
  label: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}