# lib/api-zod/src/generated/types/apiKey.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lonely Bypass API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKey {
  id: string;
  key: string;
  label: string;
  createdAt: string;
}