# lib/api-zod/src/generated/types/bulkBypassResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lonely Bypass API
 * OpenAPI spec version: 0.1.0
 */
import type { BypassResult } from './bypassResult';

export interface BulkBypassResult {
  results: BypassResult[];
}