# lib/api-zod/src/generated/types/bypassResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lonely Bypass API
 * OpenAPI spec version: 0.1.0
 */

export interface BypassResult {
  success: boolean;
  destination: string;
  /** @nullable */
  service?: string | null;
  cached?: boolean;
}