# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lonely Bypass API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: string;
  name: string;
  domains: string[];
  category: string;
}