# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lonely Bypass API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalBypassed: number;
  totalServices: number;
  uptime: string;
}